v2

+++ controllers/hostController.js
@@ -1,14 +1,17 @@
 const Home = require("../models/home");
 const Favourite = require('../models/favourite')
+const Booking = require('../models/booking')
 
 exports.getHostHomes = async (req, res, next) => {
-  const registeredHomes = await Home.find();
+  const ownerId = res.locals.userId;
+  const registeredHomes = await Home.find({ owner: ownerId });
   res.status(200).json({ homes: registeredHomes });
 };
 
 exports.getHome = async (req, res, next) => {
   const homeId = req.params.homeId;
-  const home = await Home.findById(homeId);
+  const ownerId = res.locals.userId;
+  const home = await Home.findOne({ _id: homeId, owner: ownerId });
   if (!home) {
     return res.status(404).json({ message: "Home not found" });
   }
@@ -17,7 +20,8 @@ exports.getHome = async (req, res, next) => {
 
 exports.postAddHome = async (req, res, next) => {
   const { houseName, price, location, rating, photoUrl } = req.body;
-  const home = new Home({ houseName, price, location, rating, photoUrl });
+  const owner = res.locals.userId;
+  const home = new Home({ houseName, price, location, rating, photoUrl, owner });
 
   await home.save();
 
@@ -26,9 +30,10 @@ exports.postAddHome = async (req, res, next) => {
 
 exports.putEditHome = async (req, res, next) => {
   const homeId = req.params.homeId;
+  const ownerId = res.locals.userId;
   const { houseName, price, location, rating, photoUrl } = req.body;
 
-  let home = await Home.findById(homeId);
+  let home = await Home.findOne({ _id: homeId, owner: ownerId });
   if (!home) {
     return res.status(404).json({ message: "Home not found" });
   }
@@ -44,8 +49,31 @@ exports.putEditHome = async (req, res, next) => {
 
 exports.deleteHome = async (req, res, next) => {
   const homeId = req.params.homeId;
+  const ownerId = res.locals.userId;
+
+  const home = await Home.findOne({ _id: homeId, owner: ownerId });
+  if (!home) {
+    return res.status(404).json({ message: "Home not found" });
+  }
+
   await Home.findByIdAndDelete(homeId);
   await Favourite.deleteMany({ homeId });
+  await Booking.deleteMany({ homeId });
 
   res.status(200).json({ message: "Home deleted" });
+};
+
+// Booking track record for the logged-in host's homes
+exports.getHostBookings = async (req, res, next) => {
+  const ownerId = res.locals.userId;
+
+  const myHomes = await Home.find({ owner: ownerId });
+  const myHomeIds = myHomes.map((home) => home._id);
+
+  const bookings = await Booking.find({ homeId: { $in: myHomeIds } })
+    .populate('homeId')
+    .populate('userId', 'firstName lastName email')
+    .sort({ bookedAt: -1 });
+
+  res.status(200).json({ bookings: bookings });
 };
